refactor custom script
